refactor: move api position types to `trade::position::api`
This was the last types remaining hence the `api_model` module is now removed :)
