--- mobile/native/src/api_model/mod.rs
@@ -1,1 +1,0 @@
-pub mod position;
