+++ mobile/native/src/lib.rs
@@ -1,5 +1,4 @@
 mod api;
-pub mod api_model;
 mod bridge_generated;
 pub mod calculations;
 pub mod common;
