+++ mobile/native/src/trade/position/mod.rs
@@ -1,5 +1,6 @@
 use crate::trade::ContractSymbolTrade;
 use crate::trade::DirectionTrade;
+pub mod api;
 pub mod handler;
 pub mod subscriber;
